Update helper.rb
In fix_relative_paths, prepend site.source instead of site.dest.

This eliminates the need to run jekyll twice when buiding a
PDF and a local css or image file has been updated since those
may not have been updated in _site before the PDF is generated.

Also strip off site.baseurl if it is set, otherwise PDF generation
will fail.  The baseurl is the path to the dest directory as seen
from an external web server.